Migrate several middlewares to formatters
